Move Queue related API func in its dedicated file
